Feature/day 07 (#7)
Finished day 7; pretty done with Intcode stuff.

Changed run_program function to Program class; this makes it a LOT easier to pause and unpause the runnning of the program at will. I think this can be refactored down a lot more, but I'll probably do that once I get around to doing day 9. Pretty happy with this code for now, so I'll leave it be.