Finished day 09.
The thing with these Intcode compilers is that once I have it working, it all seems to make sense. But the debugging is hell... At least we have a working Intcode computer now, so I should no longer need to worry about that. Also, I refactored day 05 to use the current version of our Intcode computer (that I now call the Program class), so day 05 and day 09 are basically identical now. Could (read: should, but probably won't) refactor day 07 as well.