Feature/day 13 (#12)
* Finished day 13, pretty clean code in the end.

Quite happy with how my Intcode program implementation worked out here;
spent some time trying to play the game myself but decided I was too
lazy to finish it, so I wrote a very very basic AI that simply follows
the ball, which turned out to be plenty for my case. I ended up not
drawing the grid in the final implementation because the game plays
itself, but I left the print_grid() function in either way. Very curious
 to see what Intcode implementations are next!